services/advertiser-api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.6.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
redis[asyncio]==5.0.1
python-jose[cryptography]==3.3.0
trino==0.327.0
celery==5.3.6
prometheus-fastapi-instrumentator==6.1.0
structlog==23.3.0
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1
bcrypt==4.1.2
